Create schema file to contain the schemas and modules for utilize in any file in views
