+++ backend/api/views/ticket.py
@@ -1,17 +1,4 @@
-from ninja import Router, Schema, NinjaAPI, Field
-from django.contrib.auth import authenticate, login
-from django.contrib.auth.models import User
-from ninja import NinjaAPI, Router, Schema, ModelSchema, Form
-from typing import List, Optional
-from api.models import *
-from django.contrib.auth.hashers import make_password
-from pydantic import field_validator
-from rest_framework_simplejwt.tokens import RefreshToken
-from datetime import datetime
-from ninja.responses import Response
-from rest_framework import status
-from django.shortcuts import get_object_or_404
-from ninja_jwt.authentication import JWTAuth
+from .schemas import *
 
 router = Router()
